// projects/employeeportal/src/app/core/models/user.ts
export interface User {

        userid: number;
        firstname: string;
        lastname: string;
        email: string;
        cellnumber: number;
        photo: string;
        birthday: Date;
        anniversary: Date;
        office: string;
        department: string;
        jobtitle: string;
        group: string;
        officedid?: string;
        officeext?: number;

}